# backend/src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AI-Library-Backend-API",
    "description": "Backend API Document"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
...
  "paths": {
...
    "/api/borrows": {
      "get": {
        "tags": ["Borrows"],
        "summary": "获取借阅记录列表",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "借阅记录列表"
          }
        }
      }
    },
    "/api/borrows/{id}": {
      "get": {
        "tags": ["Borrows"],
        "summary": "获取借阅详情",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "借阅详情"
          }
        }
      }
    },
    "/api/recommendations": {
      "get": {
        "tags": ["Recommendations"],
        "summary": "获取智能推荐",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "model",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "推荐结果"
          }
        }
      }
    },
    "/api/recommendations/history": {
      "get": {
        "tags": ["Recommendations"],
        "summary": "获取推荐历史记录",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "历史记录"
          }
        }
      }
    },
    "/api/recommendations/history/{id}": {
      "put": {
        "tags": ["Recommendations"],
        "summary": "更新推荐记录状态",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "is_rejected": {
                    "type": "boolean"
                  }
                },
                "example": {
                  "is_rejected": true
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功"
          }
        }
      }
    },
...
  }
}